library collection add and edit

+++ database/structBook.go
@@ -16,12 +16,12 @@ type Book struct {
 
 type BookDetail struct {
 	gorm.Model
-	ID          int     `json:"id,omitempty" gorm:"type:int;primaryKey;size:32;autoIncrement:false"`
-	ReleaseDate string  `json:"release_date" gorm:"type:varchar(32)"`
-	Description string  `json:"description" gorm:"type:text"`
-	Language    string  `json:"language" gorm:"type:varchar(64)"`
-	Country     string  `json:"country" gorm:"type:varchar(32)"`
-	Publisher   string  `json:"publisher" gorm:"type:varchar(32)"`
-	PageCount   float64 `json:"page_count" gorm:"type:int;size:32"`
-	Category    string  `json:"category" gorm:"type:varchar(64)"`
+	ID          int    `json:"id,omitempty" gorm:"type:int;primaryKey;size:32;autoIncrement:false"`
+	ReleaseDate string `json:"release_date" gorm:"type:varchar(32)"`
+	Description string `json:"description" gorm:"type:text"`
+	Language    string `json:"language" gorm:"type:varchar(64)"`
+	Country     string `json:"country" gorm:"type:varchar(32)"`
+	Publisher   string `json:"publisher" gorm:"type:varchar(32)"`
+	PageCount   int    `json:"page_count" gorm:"type:int;size:32"`
+	Category    string `json:"category" gorm:"type:varchar(64)"`
 }
